Bid APIs added, some functionalties are also added, such as validation for bid offers etc.

+++ backend/artus/src/main/java/com/artus/artus/models/Auction.java
@@ -16,6 +16,6 @@ public class Auction {
     private LocalDateTime start_date;
     private LocalDateTime end_date;
     private String type;
-    private float starting_amount;
+    private double starting_amount;
     private String status;
 }
